Update config.yml
Analyze Phase
UdaPeople handles some private information like social security numbers, salary amount, etc. It would be a shame if a package with a known vulnerability left a security hole in our application, giving hackers access to that information! That’s why we should include a job that checks for known vulnerabilities every time we check in new code.

Find the jobs named scan-frontend and scan-backend in the config file.
For both jobs, select a Docker image that is compatible with NodeJS.
Write code to check for security vulnerabilities in the packages used in the application.
Use npm to “audit” the code to check for known package vulnerabilities. Just cd into the directory of front-end and back-end and run the following:
npm audit --audit-level=critical